// Electron/src/types/WindowType.ts
export enum WindowType {
  MAIN = 'mainWindow',
  FOOTER = 'footerWindow',
  SETTING = 'settingWindow',
  ATTENDEE_MODAL = 'attendeeWindow',
  MEMBER = 'memberWindow',
  SCREEN = 'screenWindow',
  HOST_MODAL = 'hostWindow',
  HOST_RECORD_COMFIRM = 'hostRecordComfirm',
  HOST_RECORD_RECIVE = 'hostRecordRecive',
  APPLY_LINK = 'applyLinkWindow',
}